+++ rust/azure_iot_operations_services/src/azure_device_registry/models.rs
@@ -1,0 +1,10 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+//! Models for Azure Device Registry operations.
+
+pub mod asset;
+pub mod device;
+
+pub use asset::*;
+pub use device::*;
